Changes:- Rebased with master and changes to incorporate the enities.
